bigdecimal price fix + ids fix

+++ Application/BikeWebApp/src/main/java/com/bikeparadise/bikewebapp/dto/BikeAddDto.java
@@ -2,13 +2,14 @@ package com.bikeparadise.bikewebapp.dto;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
 
 @Getter
 public class BikeAddDto {
     String modelName;
-    Double price;
+    BigDecimal price;
     String description;
     String bikeIdentificationsAvailable;
     List<BikeAddFiltersDto> parts;
